requirements.txt
geopandas>=0.14
pandas>=2.1
numpy>=1.25
shapely>=2.0
pyproj>=3.6
h3>=3.7
h3pandas>=0.2
r5py>=0.8
gtfs-kit>=8.0
overpy>=0.7
matplotlib>=3.8
pyyaml>=6.0
requests>=2.31
